+++ android/app/src/main/java/com/stingfit/app/MainActivity.java
@@ -1,0 +1,5 @@
+package com.stingfit.app;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}